requirements.txt
# Core
mcp[cli]>=1.0.0
pydantic>=2.0.0
uvicorn[standard]>=0.30.0
fastapi>=0.111.0

# Environment config
python-dotenv>=1.0.0
